GOVSI-185: move back phone number helper as no longer used in account management
